requirements.txt
flask==3.0.3
py-cord==2.6.1  # Note: Python 3.13 compatibility issue - use Python 3.12
sqlalchemy==2.0.36
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
PyNaCl==1.5.0
qrcode==8.0
pillow==11.0.0
requests==2.32.3
pytz==2024.2
setuptools==75.6.0 